+++ doc/history.md
@@ -8,6 +8,7 @@
 - fix 解析邮件内嵌图片处理 image -> inline
 - 捕获文件上传处理异常。这个偶尔出现，但会使Server崩溃
 - 显示文件类型图标
+- 使用Meteor-Template-helpers简化模板代码
 
 ## 12-18